app/requirements.txt
sgiref==3.3.4
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-timezone-field==4.1.2
djangorestframework==3.12.4
drf-spectacular==0.15.1
factory-boy==3.2.0
Faker==8.1.1
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
jsonschema==3.2.0
oauthlib==3.1.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
